refactor(bookmarks/cron): enhance cron job handling and server-side imports
