+++ .env-example
@@ -22,6 +22,7 @@ GROQ_API_KEY=
 OPEN_API_KEY=
 BOOKMARK_BEARER_TOKEN=
 BOOKMARKS_LIST_ID=
+BOOKMARK_CRON_REFRESH_SECRET=
 GITHUB_ACCESS_TOKEN_COMMIT_GRAPH=
 
 # GitHub Activity Refresh
